apps/face-sidecar/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
insightface==0.7.3
onnxruntime==1.20.1
numpy>=1.26.0,<2.0
opencv-python-headless==4.10.0.84
psycopg2-binary==2.9.10
pgvector==0.3.6
Pillow==11.0.0
